+++ src/pages/SignIn/SignIn.scss
@@ -1,0 +1,120 @@
+.userContainer {
+  display: flex;
+  flex-shrink: 0;
+  justify-content: center;
+  width: 100%;
+  height: 100%;
+  margin-bottom: 30px;
+  .layOut {
+    width: 460px;
+    .icon {
+      display: flex;
+      // margin: 30px;
+      margin-top: 30px;
+      margin-bottom: 30px;
+    }
+    h1 {
+      margin-bottom: 40px;
+      font-weight: normal;
+      line-height: 1.2em;
+    }
+    form {
+      background-color: transparent;
+
+      .controlHeightPw {
+        flex-direction: column;
+        width: 100%;
+        input {
+          width: 460px;
+          height: 50px;
+          margin-left: 4px;
+          margin-inline-start: 9px;
+          border: 0;
+          background-color: transparent;
+          color: rgb(17, 17, 17);
+        }
+        .pw {
+          display: flex;
+          border: 1px solid #757575;
+          border-radius: 8px;
+          .pwIcon {
+            margin-top: 12px;
+            margin-inline-end: 4px;
+          }
+        }
+      }
+
+      .text {
+        margin-top: 10px;
+        .textLength8,
+        .textLength1 {
+          color: #d43f20;
+          font-size: 12px;
+        }
+      }
+
+      .date {
+        display: flex;
+        justify-content: center;
+        height: 50px;
+        margin-top: 10px;
+        margin-bottom: 10px;
+        border: 1px solid #757575;
+        border-radius: 8px;
+
+        .birthday {
+          height: 30px;
+          margin-top: 12px;
+          // width: 200px;
+          margin-left: 10px;
+          input {
+            width: 53px;
+            border: 0;
+            background-color: white;
+          }
+        }
+        .dash {
+          margin-top: 12px;
+        }
+        .lastNumber {
+          margin-top: 12px;
+          input {
+            width: 20px;
+            margin-left: 10px;
+            border: 0;
+            background-color: white;
+          }
+        }
+        .xxxxxx {
+          margin-top: 12px;
+        }
+      }
+    }
+    .btn {
+      button {
+        width: 460px;
+        height: 50px;
+        margin-top: 30px;
+        border: 0;
+        border: 1px solid black;
+        border-radius: 30px;
+        background-color: #111111;
+        color: #ffffff;
+      }
+      button:hover {
+        background-color: #757575;
+      }
+    }
+    .link {
+      margin-top: 30px;
+    }
+  }
+}
+.wrong {
+  color: #d43f20;
+  font-size: 12px;
+}
+.memberRewards {
+  color: #999;
+  font-size: 12px;
+}
